Add: Users, login, logout, register Update: style.css:

+++ hapus.php
@@ -1,6 +1,11 @@
 <?php
 include 'koneksi.php';
 
+if (!isset($_SESSION['user_id'])) {
+    header('Location: login.php');
+    exit;
+}
+
 $id = $_GET['id'];
 
 mysqli_query($conn, "DELETE FROM riwayat_gizi WHERE id=$id");
